Evaluate ini values
Since ini needs to evaluate expressions, it needs to import
the compiler, but this unfortunately has cascading import cycles.
I can't afford that a massive architectural refactor right now,
so as a workaround, I made a phpv.IniCfg interface, and
let the functions further up the call stack (such as main)
constructor the actual ini objects.