Add reference params inside go-ext functions
Ext functions have metadata comments such as:
// > func int preg_match ( string $pattern , string $subject [, array &$matches ] )

Currently, only the function name is used when generated with make buildext.
There needs a way to know if a param is a reference, and for
user-defined functions inside a PHP script, these are already available.

But this is not the case with go-ext functions. One way to do
add param info is to further parse the comment data with the buildext tool.

Although the param info given by the metadata comment
serve as a useful quick reference when writing
the implementation, for the mean time I opted not to parse and use them.
One reason is that the documentation syntax differs between versions,
and sometimes used inconsistently. The other reason is that parsing
the comments adds more work on making sure the comment is syntactically
valid.

Instead, I opted to add a go function to mark the params as a reference.