Fix wrong indices when appending to arrays
As stated in the documentation (more of a footnote really):

  The key is optional. If it is not specified, PHP will use the
  increment of the largest previously used int key.

This means that in PHP, ['a', 9=>'b','c', 2=>'d', 'e']
is actually ['a', 9=>'b', 10=>'c', 2=>'d', 11=>'e']
    and not ['a', 9=>'b', 0=>'c',  2=>'d', 1=>'e']

This is actually an unexpected behaviour,
at least compared to other languages like lua.

Anyways, this commit also fixes the bug with array_column,
array_flip, array_keys and probably a ton of others.