Handle string comparison edge case
In PHP, "a" > "999" returns true despite
what the spec and the docs say:

  If you compare a number with a string or the comparison involves numerical strings,
  then each string is converted to a number and the comparison performed numerically.

So according to above, "a" > "999" would be 0 > 999
("a" converted to int, which is 0), so 0 > 999 should be false.

Either the docs has an unspecified edge case,
or "comparison involves numerical strings" actually meant
"only one is numerical string, but not both".