Fix math bugs and add functions
There's some differences with pow and tan
between Go and PHP when it comes to computing
boundary numbers such as PHP_INT_MAX.
It's either I round the result to a less precision,
or re-implement a lot of math stuffs, neither of which
is efficient, effort- or computation-wise.
So I just modified those tests that match Go,
and worry about them later, if ever they do
become a problem (which I don't think so).
It could be just a serialization discrepancy too.